backend/pyproject.toml
[project]
name = "backend-2"
version = "0.1.0"
description = "Backend for MainData.id"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "fastapi[standard]>=0.115.6",
    "google-generativeai==0.3.2",
    "httpx==0.27.0",
    "numpy==1.26.4",
    "pgvector==0.2.5",
    "psycopg2>=2.9.10",
    "pydantic==2.6.3",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.28",
    "uvicorn==0.29.0",
]
